requirements.txt
pandas
pandas-gbq
google-auth
google-api-python-client
plotly
google-cloud-bigquery
db-dtypes
google-cloud-bigquery-storage